feat(tui): sync tree cursor to symbol when diff pane is scrolled
Per ADR 0030: manually scrolling the diff pane (j/k, mouse wheel,
half-page/top-bottom keys, ]c/[c) while Focus::Right now moves the
tree cursor onto whichever symbol section the new scroll offset falls
inside, mirroring ADR 0027's existing tree -> diff auto-scroll in the
opposite direction.

- diff_shape::symbol_id_for_scroll_line: new pure reverse lookup
  (start_line -> symbol_id), sitting next to section_start_line_for_symbol
  (the forward lookup) and reusing the same walk_sections layout table.
  Returns None for the module-level bucket or an overscroll past the
  last section - the tree cursor is left untouched in both cases
  rather than guessing a nearest symbol.
- App::sync_tree_cursor_to_symbol: moves the cursor via the existing
  Nav::move_cursor_to_symbol (expanding collapsed ancestors), but
  deliberately not App::jump_to_symbol - no jumplist entry, no
  right_pane_scroll reset, since both would fight this sync's own
  trigger.
- run_app's per-key diff-pane-content recompute + both sync directions
  is extracted into a new pure apply_diff_pane_selection_effects
  function (mirroring why dispatch_non_source_key exists as its own
  function), so the feedback-loop guard - last_diff_focus must reflect
  the post-sync symbol before the next handled key runs, or the next
  key's auto-scroll would snap right_pane_scroll straight back - has
  direct regression coverage instead of only code review.

Moves lib.rs's test module out to a sibling tests.rs
(#[cfg(test)] #[path = "tests.rs"] mod tests;), the same pattern
CLAUDE.md's file-size discipline documents and PR #82 already
established, since lib.rs had already crossed the warn threshold on
test-code weight and this change's own new tests would have pushed it
further.
